requirements.txt
pydantic-ai==0.0.15
polygon-api-client==1.14.2